change guess heuristic for non-bivalue cases
most of the time a cell with exactly 2 possibilities exist
in the rare case that it doesn't, the code used to just take the first
cell and go through all its possibilities.
This results in suboptimal guess paths because the cell likely contains
many possibilities. On the other hand, looking through all unsolved
cells for the minimum of possibilities is expensive, too.
As a compromise, the code now looks through up to 3 cells, one for each
of the digit sets 1-3, 4-6, 7-9

The change affects only harder sudokus or impossible ones.