clarify and simplify the update routine
row_uniq is identical to shrink_mask after the inversions of the last
commit and is removed

the modifications of unsolved_rows (ur ^= s << shift) turned out to be
both bogus, but also a harmless mistake as the solver
will just turn to trial and error instead. It exists purely for optimization
to avoid unnecessary work. However, it actually slows things down.
Removed unsolved rows bookkeeping

because updn_upwcl was called unconditionally, shrink may be set to 0
again after making new deductions, so it is no longer a good flag for
breaking the loop. Replaced by boolean flag.