overhaul position system for strategy solver
And change the strategy solver to use it. The generator
still uses the old system.

This adds types for various concepts as well as conversion
functions between them. Bitmasks are now available under the
Set<T> type and differentiate between Digit, Position<Row>
and Position<Col>, etc.

This is just the first step, the old position system will be removed
and the new one needs to be cleaned up