commit Cargo.lock
compilation on the previous target of 1.51 failed because there was no
Cargo.lock. Newer versions of some dependencies
raised their MSRV and so CI would get those versions and fail to compile
the crate.
This wouldn't have happened when cargo would allow specifiying rust
versions, but until that's possible, Cargo.lock it is.